backend/requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bcrypt==3.2.2
billiard==3.6.4.0
bitarray==2.6.0
boto3==1.18.51
botocore==1.21.65
CacheControl==0.12.11
cachetools==5.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codeguru-profiler-agent==1.2.4
coverage==6.4.3
cryptography==37.0.4
cytoolz==0.12.0
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
elastic-apm==6.12.0
elasticsearch==7.13.0
email-validator==2.0.0.post1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastapi==0.75.0
fastapi-cache2==0.1.9
fastapi-cprofile==0.0.2
fastapi-redis-cache==0.2.5
fastapi-utils==0.2.1
filelock==3.0.12
firebase-admin==6.0.1
frozenlist==1.3.1
gcloud==0.18.3
google-api-core==2.10.1
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
greenlet==1.1.0
grpcio==1.54.0
grpcio-status==1.49.0rc1
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.3
hiredis==2.0.0
httplib2==0.22.0
httptools==0.1.1
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
jmespath==0.10.0
jsonschema==4.9.1
jwcrypto==1.4.2
kombu==5.2.4
lru-dict==1.1.8
msgpack==1.0.5
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauth2client==4.1.3
openai==0.27.4
packaging==23.1
parsimonious==0.8.1
passlib==1.7.4
pbr==5.10.0
pendulum==2.1.2
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.30
proto-plus==1.22.2
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
PyJWT==2.6.0
pyparsing==2.4.7
Pyrebase4==4.6.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-json-logger==2.0.1
python-jwt==4.0.0
python-multipart==0.0.5
pytz==2022.2
pytzdata==2020.1
redis==4.4.0rc1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.10.1
rlp==2.0.1
rsa==4.9
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
starlette==0.17.1
Tempita==0.5.2
toolz==0.12.0
tqdm==4.65.0
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.13.4
uvloop==0.15.2
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.4.0
